# This is a combination of 2 commits. # This is the 1st commit message:
Added tests

# This is the commit message #2:

# This is a combination of 2 commits.
# This is the 1st commit message:

Added tests

# This is the commit message #2:

# This is a combination of 51 commits.
# This is the 1st commit message:

Initial commit

Gitignore updated

Improved architecture

Base package name refactored

Made it work

Adding tests

# This is the commit message #2:

Adding tests

# This is the commit message #3:

Added file logger

# This is the commit message #4:

Added encrypted properties

# This is the commit message #5:

Gitignore updated

# This is the commit message #6:

Demo borderless

# This is the commit message #7:

Demo borderless

# This is the commit message #8:

Refactoring

# This is the commit message #9:

Scroll bars and table styled

# This is the commit message #10:

Adjuseted styles

# This is the commit message #11:

Dialog button

# This is the commit message #12:

Major progress with dialog

# This is the commit message #13:

Working fully implemented dialog

# This is the commit message #14:

DB changed to PostgerSQL

# This is the commit message #15:

Implemented cmd bat call

# This is the commit message #16:

External user change lock

# This is the commit message #17:

Started styling

# This is the commit message #18:

Merge cleanup

# This is the commit message #19:

Refactoring

# This is the commit message #20:

Refactoring

# This is the commit message #21:

Side bar menu and font awesome icons

# This is the commit message #22:

Experimenting with font and design

# This is the commit message #23:

Experimenting with font and design

# This is the commit message #24:

Adding customers

# This is the commit message #25:

Styled window buttons

# This is the commit message #26:

BugFix

# This is the commit message #27:

Labeled row

# This is the commit message #28:

Labeled row

# This is the commit message #29:

Addedarticles

# This is the commit message #30:

Bugfix

# This is the commit message #31:

Added city and sanitise period to customer

# This is the commit message #32:

Autocomplete city in customer dialog

# This is the commit message #33:

Maximize Normalize button fixed

# This is the commit message #34:

Article tables dynamically create columns, Customer and Article ID changed type, additional tables created in db.

# This is the commit message #35:

Added mandatory articles

# This is the commit message #36:

WIP - Orders

# This is the commit message #37:

WIP - Orders

# This is the commit message #38:

Odrers WIP 2

# This is the commit message #39:

Read from db implemented

# This is the commit message #40:

Read from db implemented improved

# This is the commit message #41:

Total price added

# This is the commit message #42:

Orders completed - (need to add validations)

# This is the commit message #43:

Refactoring

# This is the commit message #44:

Refactoring

# This is the commit message #45:

Date picker formatting

# This is the commit message #46:

Adding logic for tracking packages and debt

# This is the commit message #47:

Adding logic for tracking packages and debt bugfix

# This is the commit message #48:

Adding logic for tracking packages and debt improved logic

# This is the commit message #49:

Small styling changes

# This is the commit message #50:

Performance and style improvement

# This is the commit message #51:

Style improvement

+++ src/main/java/com/nautilus/service/impl/OrderServiceImpl.java
@@ -12,8 +12,6 @@ import com.nautilus.util.Formatter;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.annotation.PostConstruct;
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -30,11 +28,6 @@ public class OrderServiceImpl implements OrderService {
 
     final
     CustomerService customerService;
-
-    @PostConstruct
-    public void init() {
-        customerService.setOrderService(this);
-    }
 
     public OrderServiceImpl(OrderItemService orderItemService, OrderRepository orderRepository, CustomerService customerService) {
         this.orderItemService = orderItemService;
@@ -54,8 +47,9 @@ public class OrderServiceImpl implements OrderService {
         List<OrderItem> unsavedItems = obj.getItems();
         obj.clearItems();
         unsavedItems.forEach(item -> savedOrder.addItem(orderItemService.insert(item)));
-        updateCustomersObligations(savedOrder.getCustomer(), savedOrder);
-        customerService.update(savedOrder.getCustomer());
+        updateCustomersObligations(obj.getCustomer(), obj);
+        updateCustomersPackagingFlags(obj.getCustomer());
+        customerService.update(obj.getCustomer());
         return savedOrder;
     }
 
@@ -67,15 +61,8 @@ public class OrderServiceImpl implements OrderService {
     }
 
     @Override
-    public List<OrderDto> getAllDto(LocalDate date, String city) {
-        return orderRepository.getAllDto(date, city);
-    }
-
-    @Override
-    public Optional<Order> getCustomersLastOrder(Customer customer) {
-        Optional<Order> customersLastOrder = orderRepository.getCustomersLastOrder(customer);
-        customersLastOrder.ifPresent(order -> orderItemService.getAllByOrderId(order.getId()).forEach(order::addItem));
-        return customersLastOrder;
+    public List<OrderDto> getAllDto() {
+        return orderRepository.getAllDto();
     }
 
     @Override
@@ -88,15 +75,18 @@ public class OrderServiceImpl implements OrderService {
     @Override
     public Optional<Order> update(Order obj) {
         Order oldOrder = this.findById(obj.getId());
-        orderRepository.update(obj);
-        List<OrderItem> unsavedItems = obj.getItems();
-        obj.clearItems();
-        orderItemService.deleteAllByOrderId(obj.getId());
-        unsavedItems.forEach(orderItem -> obj.addItem(orderItemService.insert(orderItem)));
-        revertCustomersObligations(obj.getCustomer(), oldOrder);
-        updateCustomersObligations(obj.getCustomer(), obj);
-        customerService.update(obj.getCustomer());
-        return Optional.of(obj);
+        Optional<Order> updatedOrder = orderRepository.update(obj);
+        if (updatedOrder.isPresent()) {
+            List<OrderItem> unsavedItems = obj.getItems();
+            updatedOrder.get().clearItems();
+            orderItemService.deleteAllByOrderId(obj.getId());
+            unsavedItems.forEach(orderItem -> updatedOrder.get().addItem(orderItemService.insert(orderItem)));
+            revertCustomersObligations(updatedOrder.get().getCustomer(), oldOrder);
+            updateCustomersObligations(updatedOrder.get().getCustomer(), updatedOrder.get());
+            updateCustomersPackagingFlags(obj.getCustomer());
+            customerService.update(obj.getCustomer());
+        }
+        return updatedOrder;
     }
 
     @Override
@@ -105,6 +95,7 @@ public class OrderServiceImpl implements OrderService {
             Order order = this.findById(orderDto.getId());
             orderRepository.deleteById(order);
             revertCustomersObligations(order.getCustomer(), order);
+            updateCustomersPackagingFlags(order.getCustomer());
             customerService.update(order.getCustomer());
         });
     }
@@ -112,12 +103,13 @@ public class OrderServiceImpl implements OrderService {
     @Override
     public void deleteAllDto(Iterable<OrderDto> obj) {
         List<Order> orders = new ArrayList<>();
-        obj.forEach(orderDto -> {
-            if (orderDto.getDeliveredBy().equals(Order.DeliveredBy.NONE.toString())) {
-                orders.add(mapToOrder(orderDto));
-            }
-        });
+        obj.forEach(orderDto -> orders.add(mapToOrder(orderDto)));
         this.deleteAll(orders);
+    }
+
+    @Override
+    public List<Order> getAllByCustomerIdOrderByDate(Long customerId) {
+        return orderRepository.getAllByCustomerIdOrderByDate(customerId);
     }
 
     private void updateCustomersObligations(Customer customer, Order order) {
@@ -127,7 +119,7 @@ public class OrderServiceImpl implements OrderService {
 
         }
         if (!order.getDeliveredBy().equals(Order.DeliveredBy.NONE) && !order.getPayed()) {
-            customer.setDebt(customer.getDebt() + ((Double) order.getItems().stream().map(orderItem -> orderItem.getArticlePrice() * (1 + orderItem.getArticleTax() / 100) * orderItem.getQuantity()).mapToDouble(Double::doubleValue).sum()));
+            customer.setDebt(customer.getDebt() + ((Double) order.getItems().stream().map(orderItem -> orderItem.getArticlePrice() * orderItem.getQuantity()).mapToDouble(Double::doubleValue).sum()));
         }
     }
 
@@ -137,7 +129,15 @@ public class OrderServiceImpl implements OrderService {
             customer.setPackagingLarge(customer.getPackagingLarge() - order.findOrderItemByArticleName(Formatter.WATER_LARGE).map(OrderItem::getQuantity).orElse(0));
         }
         if (!order.getDeliveredBy().equals(Order.DeliveredBy.NONE) && !order.getPayed()) {
-            customer.setDebt(customer.getDebt() - ((Double) order.getItems().stream().map(orderItem -> orderItem.getArticlePrice() * (1 + orderItem.getArticleTax() / 100) * orderItem.getQuantity()).mapToDouble(Double::doubleValue).sum()));
+            customer.setDebt(customer.getDebt() - ((Double) order.getItems().stream().map(orderItem -> orderItem.getArticlePrice() * orderItem.getQuantity()).mapToDouble(Double::doubleValue).sum()));
+        }
+    }
+
+    private void updateCustomersPackagingFlags(Customer customer) {
+        Optional<Order> lastOrder = orderRepository.getCustomersLastOrder(customer.getId());
+        if (lastOrder.isPresent()) {
+            customer.setBacklogPackagingSmall(customer.getPackagingSmall() > lastOrder.get().getPackagingDebtSmall());
+            customer.setBacklogPackagingLarge(customer.getPackagingLarge() > lastOrder.get().getPackagingDebtLarge());
         }
     }
